# docs/android-handoff/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SwaggerExport",
    "version": "1.0"
  },
  "paths": {
    "/api/CheckIn/A_011_CHECKIN_DS_KH_CAT_NUOC": {
      "post": {
        "tags": [
          "CheckIn"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/A_011_CHECKIN_DS_KH_CAT_NUOCRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/A_011_CHECKIN_DS_KH_CAT_NUOCRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/A_011_CHECKIN_DS_KH_CAT_NUOCRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CheckIn/A_012_CHECKIN_DS_KH_MO_NUOC": {
      "post": {
        "tags": [
          "CheckIn"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/A_012_CHECKIN_DS_KH_MO_NUOCRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/A_012_CHECKIN_DS_KH_MO_NUOCRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/A_012_CHECKIN_DS_KH_MO_NUOCRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CheckIn/A_013_CHECKIN_DS_KH_GUI_GB": {
      "post": {
        "tags": [
          "CheckIn"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/A_013_CHECKIN_DS_KH_GUI_GBRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/A_013_CHECKIN_DS_KH_GUI_GBRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/A_013_CHECKIN_DS_KH_GUI_GBRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CheckIn/A_99_CHECKIN_DM_KIEU_CAT": {
      "post": {
        "tags": [
          "CheckIn"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/A_99_CHECKIN_DM_KIEU_CATRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/A_99_CHECKIN_DM_KIEU_CATRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/A_99_CHECKIN_DM_KIEU_CATRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CheckIn/A_02_CHECKIN_LUU_KH_CAT_NUOC": {
      "post": {
        "tags": [
          "CheckIn"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/A_02_CHECKIN_LUU_KH_CAT_NUOCRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/A_02_CHECKIN_LUU_KH_CAT_NUOCRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/A_02_CHECKIN_LUU_KH_CAT_NUOCRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CheckIn/A_03_CHECKIN_LUU_KH_MO_NUOC": {
      "post": {
        "tags": [
          "CheckIn"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/A_03_CHECKIN_LUU_KH_MO_NUOCRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/A_03_CHECKIN_LUU_KH_MO_NUOCRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/A_03_CHECKIN_LUU_KH_MO_NUOCRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CheckIn/A_00_CHECKIN_LUU_TEN_FILE_ANH": {
      "post": {
        "tags": [
          "CheckIn"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/A_00_CHECKIN_LUU_TEN_FILE_ANHRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/A_00_CHECKIN_LUU_TEN_FILE_ANHRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/A_00_CHECKIN_LUU_TEN_FILE_ANHRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "A_00_CHECKIN_LUU_TEN_FILE_ANHRequest": {
        "type": "object",
        "properties": {
          "ID_XAC_NHAN": {
            "type": "string",
            "nullable": true
          },
          "LOAI_CV": {
            "type": "string",
            "nullable": true
          },
          "TEN_FILE_ANH": {
            "type": "string",
            "nullable": true
          },
          "MA_NHAN_VIEN": {
            "type": "string",
            "nullable": true
          },
          "SO_IMEI": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_K": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "A_011_CHECKIN_DS_KH_CAT_NUOCRequest": {
        "type": "object",
        "properties": {
          "MA_NHAN_VIEN": {
            "type": "string",
            "nullable": true
          },
          "TU_NGAY": {
            "type": "string",
            "nullable": true
          },
          "DEN_NGAY": {
            "type": "string",
            "nullable": true
          },
          "MA_XI_NGHIEP": {
            "type": "string",
            "nullable": true
          },
          "SO_IMEI": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_K": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "A_012_CHECKIN_DS_KH_MO_NUOCRequest": {
        "type": "object",
        "properties": {
          "MA_NHAN_VIEN": {
            "type": "string",
            "nullable": true
          },
          "TU_NGAY": {
            "type": "string",
            "nullable": true
          },
          "DEN_NGAY": {
            "type": "string",
            "nullable": true
          },
          "MA_XI_NGHIEP": {
            "type": "string",
            "nullable": true
          },
          "SO_IMEI": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_K": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "A_013_CHECKIN_DS_KH_GUI_GBRequest": {
        "type": "object",
        "properties": {
          "MA_NHAN_VIEN": {
            "type": "string",
            "nullable": true
          },
          "TU_NGAY": {
            "type": "string",
            "nullable": true
          },
          "DEN_NGAY": {
            "type": "string",
            "nullable": true
          },
          "MA_XI_NGHIEP": {
            "type": "string",
            "nullable": true
          },
          "SO_IMEI": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_K": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "A_02_CHECKIN_LUU_KH_CAT_NUOCRequest": {
        "type": "object",
        "properties": {
          "ID_XAC_NHAN": {
            "type": "string",
            "nullable": true
          },
          "MA_NHAN_VIEN": {
            "type": "string",
            "nullable": true
          },
          "MA_XI_NGHIEP": {
            "type": "string",
            "nullable": true
          },
          "NGUOI_THI_CONG": {
            "type": "string",
            "nullable": true
          },
          "MA_KIEU_CAT_MO": {
            "type": "string",
            "nullable": true
          },
          "NGAY_HOAN_THANH": {
            "type": "string",
            "nullable": true
          },
          "VI_TRI_XAC_NHAN": {
            "type": "string",
            "nullable": true
          },
          "MA_GHI_CHU": {
            "type": "string",
            "nullable": true
          },
          "GHI_CHU": {
            "type": "string",
            "nullable": true
          },
          "SO_IMEI": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_K": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "A_03_CHECKIN_LUU_KH_MO_NUOCRequest": {
        "type": "object",
        "properties": {
          "ID_XAC_NHAN": {
            "type": "string",
            "nullable": true
          },
          "MA_NHAN_VIEN": {
            "type": "string",
            "nullable": true
          },
          "MA_XI_NGHIEP": {
            "type": "string",
            "nullable": true
          },
          "NGUOI_THI_CONG": {
            "type": "string",
            "nullable": true
          },
          "MA_KIEU_CAT_MO": {
            "type": "string",
            "nullable": true
          },
          "NGAY_HOAN_THANH": {
            "type": "string",
            "nullable": true
          },
          "VI_TRI_XAC_NHAN": {
            "type": "string",
            "nullable": true
          },
          "MA_GHI_CHU": {
            "type": "string",
            "nullable": true
          },
          "GHI_CHU": {
            "type": "string",
            "nullable": true
          },
          "SO_IMEI": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_K": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "A_99_CHECKIN_DM_KIEU_CATRequest": {
        "type": "object",
        "properties": {
          "MA_XI_NGHIEP": {
            "type": "string",
            "nullable": true
          },
          "SO_IMEI": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_K": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}